requirements.txt
pre-commit==3.5.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-github-actions-annotate-failures==0.3.0
pydantic==2.6.4
pyyaml==6.0.1
requests==2.32.2
pdfplumber==0.10.2
pytesseract==0.3.10
Pillow==10.2.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
jsonschema==4.21.1
pykwalify==1.8.0
fastapi==0.110.0
uvicorn==0.27.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
langchain==0.1.15
langgraph==0.0.24
langsmith==0.1.18
weaviate-client==3.26.7
# Optional heavy dependencies removed for tests

